Initialize ids on connect and connection data. Add protectedids storage handler
