+++ package.json
@@ -39,7 +39,7 @@
     "@material-ui/icons": "^4.11.2",
     "@psychedelic/dab-js": "1.2.1",
     "@psychedelic/plug-controller": "0.15.5",
-    "@psychedelic/plug-inpage-provider": "1.8.0",
+    "@psychedelic/plug-inpage-provider": "1.8.1",
     "@reduxjs/toolkit": "^1.6.0",
     "advanced-css-reset": "^1.2.2",
     "axios": "^0.21.1",
