+++ source/Background/utils.js
@@ -1,6 +1,9 @@
 import { validateCanisterId, validatePrincipalId, validateAccountId } from '@shared/utils/ids';
-import { XTC_FEE } from '@shared/constants/addresses';
+import { getDabNfts, getDabTokens } from '@shared/services/DAB';
+import { ASSET_CANISTER_IDS } from '@shared/constants/canisters';
 import { CYCLES_PER_TC } from '@shared/constants/currencies';
+import { XTC_FEE } from '@shared/constants/addresses';
+import { setProtectedIds } from '@modules';
 
 import ERRORS from './errors';
 
@@ -50,3 +53,14 @@ export const validateTransactions = (transactions) => Array.isArray(transactions
   && transactions?.every(
     (tx) => tx.idl && tx.canisterId && tx.methodName && tx.args,
   );
+
+export const initializeProtectedIds = async () => {
+  const nftCanisters = await getDabNfts();
+  const tokenCanisters = await getDabTokens();
+  const PROTECTED_IDS = [
+    ...(nftCanisters || []).map((collection) => collection.principal_id.toString()),
+    ...(tokenCanisters || []).map((token) => token.principal_id.toString()),
+    ...ASSET_CANISTER_IDS,
+  ];
+  setProtectedIds(PROTECTED_IDS);
+};
